Move the docker_plugin library to a custom resource Make the defaults nil rather than empty Hashes and Arrays
